Remove test_autolog_extra_tags.py, move tests to corresponding files to fix test failures
